# app/frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/evidence/upload": {
      "post": {
        "description": "Encrypts and stores a single evidence file locally for eventual upload. Accepts one file (max 10MB) in the \"file\" field. Allowed types: image/jpeg, image/png, image/gif, application/pdf, text/plain.",
        "operationId": "EvidenceController_upload_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evidence queued successfully."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Upload evidence to queue",
        "tags": [
          "Evidence Queue"
        ]
      }
    },
    "/api/v1/evidence/queue": {
      "get": {
        "description": "Retrieves all evidence items in the queue for the current user.",
        "operationId": "EvidenceController_getQueue_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items to return (default: 25, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor for pagination",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Object"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List evidence queue",
        "tags": [
          "Evidence Queue"
        ]
      }
    },
    "/api/v1/evidence/queue/{id}/retry": {
      "post": {
        "description": "Manually triggers a retry for a failed evidence upload.",
        "operationId": "EvidenceController_retry_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retry initiated."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Retry evidence upload",
        "tags": [
          "Evidence Queue"
        ]
      }
    },
    "/api/v1/evidence/queue/{id}": {
      "delete": {
        "description": "Removes an evidence item from the queue and deletes the local file.",
        "operationId": "EvidenceController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed successfully."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remove from queue",
        "tags": [
          "Evidence Queue"
        ]
      }
    },
...
  },
  "info": {
    "title": "ChainForge API",
    "description": "API documentation for ChainForge platform - Humanitarian aid and verification system\n\n## API Versioning\n\nThis API uses URI-based versioning. The current version is **v1**.\n\n### Version Format\nAll endpoints are prefixed with the version number: `/api/v1/...`\n\n### Supported Versions\n| Version | Status | Description |\n|---------|--------|-------------|\n| v1 | Current | Active version with full support |\n\n### Deprecation Policy\n- Deprecated endpoints will be marked with `@Deprecated` in the documentation\n- Deprecated versions will be supported for at least 6 months after deprecation notice\n- Clients will receive deprecation warnings via the `Sunset` HTTP header\n- Migration guides will be provided for major version changes\n\n### Future Versions\nWhen new versions are released:\n- New endpoints will be available at `/api/v2/...`, etc.\n- Previous versions remain accessible during the deprecation period\n- Clients should monitor the API documentation for version updates",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Staging"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Enter JWT token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for external access"
      }
    },
    "schemas": {
...
      "PaginatedResult": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The array of records for the current page.",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Cursor to retrieve the next page of results, or null if there is no next page."
          }
        },
        "required": [
          "data"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}